Add testing module that exposes symbol test utils
