+++ tests/test_symbols.py
@@ -12,19 +12,18 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-import abc
 import collections.abc
 import itertools
 import math
 import operator
 import typing
 import unittest
-import warnings
 
 import numpy as np
 
 import dwave.optimization
 import dwave.optimization.symbols
+from dwave.optimization.testing import symbol_testing_utils
 from dwave.optimization import (
     Model,
     logical,
@@ -40,259 +39,7 @@ from dwave.optimization import (
 )
 
 
-class utils:
-    """We want to add subclasses of unittest.TestCase that implement universal
-    tests for different symbols. But they get executed as tests if they
-    are in the main namspace. So we nest them in this "namespace".
-    """
-    class SymbolTests(abc.ABC, unittest.TestCase):
-        @abc.abstractmethod
-        def generate_symbols(self):
-            """Yield symbol(s) for testing.
-
-            The model must be topologically sorted before returning.
-            The symbols must all be unique from eachother.
-            """
-
-        def test_equality(self):
-            DEFINITELY = 2
-            for x in self.generate_symbols():
-                self.assertEqual(DEFINITELY, x.maybe_equals(x))
-                self.assertTrue(x.equals(x))
-
-            for x, y in zip(self.generate_symbols(), self.generate_symbols()):
-                self.assertTrue(DEFINITELY, x.maybe_equals(y))
-                self.assertTrue(x.equals(y))
-
-        def test_inequality(self):
-            MAYBE = 1
-            for x, y in itertools.combinations(self.generate_symbols(), 2):
-                self.assertLessEqual(x.maybe_equals(y), MAYBE)
-                self.assertFalse(x.equals(y))
-
-        def test_iter_symbols(self):
-            for x in self.generate_symbols():
-                model = x.model
-                index = x.topological_index()
-
-                # Get the symbol back
-                y, = itertools.islice(model.iter_symbols(), index, index+1)
-
-                self.assertTrue(x.shares_memory(y))
-                self.assertIs(type(x), type(y))
-                self.assertTrue(x.equals(y))
-
-        def test_namespace(self):
-            x = next(self.generate_symbols())
-            self.assertIn(type(x).__name__, dwave.optimization.symbols.__all__)
-
-        def test_serialization(self):
-            for x in self.generate_symbols():
-                model = x.model
-                index = x.topological_index()
-
-                with model.to_file() as f:
-                    new = Model.from_file(f)
-
-                # Get the symbol back
-                y, = itertools.islice(new.iter_symbols(), index, index+1)
-
-                self.assertFalse(x.shares_memory(y))
-                self.assertIs(type(x), type(y))
-                self.assertTrue(x.equals(y))
-
-        def test_state_serialization(self):
-            for x in self.generate_symbols():
-                model = x.model
-
-                # without some way to randomly initialize states this is really just
-                # smoke test
-                model.states.resize(1)
-
-                # get the states before serialization
-                states = []
-                for sym in model.iter_symbols():
-                    if hasattr(sym, "state"):
-                        states.append(sym.state())
-                    else:
-                        states.append(None)
-
-                with model.states.to_file() as f:
-                    new = Model.from_file(f)
-
-                for i, sym in enumerate(model.iter_symbols()):
-                    if hasattr(sym, "state"):
-                        np.testing.assert_equal(sym.state(), states[i])
-
-        def test_state_size_smoke(self):
-            for x in self.generate_symbols():
-                self.assertGreaterEqual(x.state_size(), 0)
-
-    class BinaryOpTests(SymbolTests):
-        @abc.abstractmethod
-        def op(self, lhs, rhs):
-            pass
-
-        def symbol_op(self, lhs, rhs):
-            # if the op is different for symbols, allow the override
-            return self.op(lhs, rhs)
-
-        def test_numpy_equivalence(self):
-            lhs_array = np.arange(10)
-            rhs_array = np.arange(1, 11)
-
-            model = Model()
-            lhs_symbol = model.constant(lhs_array)
-            rhs_symbol = model.constant(rhs_array)
-
-            op_array = self.op(lhs_array, rhs_array)
-            op_symbol = self.symbol_op(lhs_symbol, rhs_symbol)
-
-            model.states.resize(1)
-            with model.lock():
-                np.testing.assert_array_equal(op_array, op_symbol.state())
-
-        def test_scalar_broadcasting(self):
-            lhs_array = 5
-            rhs_array = np.asarray([-10, 100, 16])
-
-            model = Model()
-            lhs_symbol = model.constant(lhs_array)
-            rhs_symbol = model.constant(rhs_array)
-
-            op_array = self.op(lhs_array, rhs_array)
-            op_symbol = self.symbol_op(lhs_symbol, rhs_symbol)
-
-            model.states.resize(1)
-            with model.lock():
-                np.testing.assert_array_equal(op_array, op_symbol.state())
-
-        def test_size1_broadcasting(self):
-            lhs_array = np.asarray([5])
-            rhs_array = np.asarray([-10, 100, 16])
-
-            model = Model()
-            lhs_symbol = model.constant(lhs_array)
-            rhs_symbol = model.constant(rhs_array)
-
-            op_array = self.op(lhs_array, rhs_array)
-            op_symbol = self.symbol_op(lhs_symbol, rhs_symbol)
-
-            model.states.resize(1)
-            with model.lock():
-                np.testing.assert_array_equal(op_array, op_symbol.state())
-
-    class NaryOpTests(SymbolTests):
-        @abc.abstractmethod
-        def op(self, *xs):
-            pass
-
-        @abc.abstractmethod
-        def node_op(self, *xs):
-            pass
-
-        @abc.abstractmethod
-        def node_class(self):
-            pass
-
-        def generate_symbols(self):
-            model = Model()
-            a, b, c = model.constant(-5), model.constant(7), model.constant(0)
-            op_abc = self.node_op(a, b, c)
-            model.lock()
-            yield op_abc
-
-        def test_scalar_input(self):
-            model = Model()
-            a, b, c = model.constant(-5), model.constant(7), model.constant(0)
-            op_abc = self.node_op(a, b, c)
-            self.assertEqual(model.num_symbols(), 4)
-
-            # Make sure we got the right type of node
-            self.assertIsInstance(op_abc, self.node_class())
-
-            # Should be a scalar
-            self.assertEqual(op_abc.shape(), ())
-            self.assertEqual(op_abc.size(), 1)
-
-            model.lock()
-            model.states.resize(1)
-            self.assertEqual(op_abc.state(0), self.op(-5, 7, 0))
-
-        def test_1d_input(self):
-            model = Model()
-            x, y, z = [model.integer(10, -5, 5) for _ in range(3)]
-            op_xyz = self.node_op(x, y, z)
-
-            # Make sure we got the right type of node
-            self.assertIsInstance(op_xyz, self.node_class())
-
-            # Make sure the shape is correct
-            self.assertEqual(op_xyz.shape(), (10,))
-            self.assertEqual(op_xyz.size(), 10)
-
-            model.lock()
-
-            model.states.resize(1)
-
-            data_x = [-5, -4, -3, -2, -1, 0, 1, 2, 3, 4]
-            x.set_state(0, data_x)
-            data_y = [5, 4, 3, 2, 1, 0, -1, -2, -3, -4]
-            y.set_state(0, data_y)
-            data_z = [5, -5, 4, -4, 3, -3, 2, -2, 1, 0]
-            z.set_state(0, data_z)
-
-            np.testing.assert_equal(
-                op_xyz.state(0),
-                [self.op(*vs) for vs in zip(data_x, data_y, data_z)]
-            )
-
-    class UnaryOpTests(SymbolTests):
-        @abc.abstractmethod
-        def op(self, x):
-            pass
-
-        def symbol_op(self, x):
-            # if the op is different for symbols, allow the override
-            return self.op(x)
-
-        def generate_symbols(self):
-            model = Model()
-            a = model.constant(-5)
-            op_a = self.symbol_op(a)
-            model.lock()
-            yield op_a
-
-        def test_scalar_input(self):
-            for scalar in [-5, -.5, 0, 1, 1.5]:
-                with self.subTest(f"a = {scalar}"):
-                    model = Model()
-                    a = model.constant(scalar)
-                    op_a = self.symbol_op(a)
-                    self.assertEqual(model.num_symbols(), 2)
-
-                    # Should be a scalar
-                    self.assertEqual(op_a.shape(), ())
-                    self.assertEqual(op_a.size(), 1)
-
-                    model.lock()
-                    model.states.resize(1)
-                    self.assertEqual(op_a.state(0), self.op(scalar))
-
-        def test_1d_input(self):
-            model = Model()
-            x = model.integer(10, -5, 5)
-            op_x = self.symbol_op(x)
-            model.lock()
-
-            model.states.resize(1)
-            data = [-5, -4, -3, -2, -1, 0, 1, 2, 3, 4]
-            x.set_state(0, data)
-
-            np.testing.assert_equal(op_x.state(0), [self.op(v) for v in data])
-
-
-class TestAbsolute(utils.UnaryOpTests):
+class TestAbsolute(symbol_testing_utils.UnaryOpTests):
     def op(self, x):
         return abs(x)
 
@@ -306,7 +53,7 @@ class TestAbsolute(utils.UnaryOpTests):
         self.assertEqual(model.num_symbols(), 2)
 
 
-class TestAdd(utils.BinaryOpTests):
+class TestAdd(symbol_testing_utils.BinaryOpTests):
     def generate_symbols(self):
         model = Model()
         a = model.constant(5)
@@ -423,7 +170,7 @@ class TestAdvancedIndexing(unittest.TestCase):
             constant[i2]
 
 
-class TestAll(utils.SymbolTests):
+class TestAll(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         nodes = [
@@ -478,7 +225,7 @@ class TestAll(utils.SymbolTests):
                 model.unlock()
 
 
-class TestAnd(utils.BinaryOpTests):
+class TestAnd(symbol_testing_utils.BinaryOpTests):
     def generate_symbols(self):
         model = Model()
         a = model.constant(1)
@@ -521,12 +268,12 @@ class TestAnd(utils.BinaryOpTests):
         self.assertEqual(cb.state(0), 0)
 
         with self.assertRaises(TypeError):
-            x = a & b
+            a & b
         with self.assertRaises(AttributeError):
-            x = a.logical_and(b)
+            a.logical_and(b)
 
 
-class TestAny(utils.SymbolTests):
+class TestAny(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         nodes = [
@@ -581,7 +328,7 @@ class TestAny(utils.SymbolTests):
                 model.unlock()
 
 
-class TestArrayValidation(utils.SymbolTests):
+class TestArrayValidation(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         x = model.binary(10)
@@ -593,7 +340,7 @@ class TestArrayValidation(utils.SymbolTests):
         pass
 
 
-class TestBasicIndexing(utils.SymbolTests):
+class TestBasicIndexing(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         symbols = []
 
@@ -673,7 +420,7 @@ class TestBasicIndexing(utils.SymbolTests):
         self.assertEqual(model.set(10)[::2].state_size(), 5 * 8)
 
 
-class TestBinaryVariable(utils.SymbolTests):
+class TestBinaryVariable(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         x = model.binary(10)
@@ -690,7 +437,7 @@ class TestBinaryVariable(utils.SymbolTests):
     def test(self):
         model = Model()
 
-        x = model.binary([10])
+        model.binary([10])
 
     def test_no_shape(self):
         model = Model()
@@ -771,13 +518,13 @@ class TestBinaryVariable(utils.SymbolTests):
                 x.set_state(0, [0, 1, 2])
 
 
-class TestConcatenate(utils.SymbolTests):
+class TestConcatenate(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
-        x = model.constant(np.arange(12)).reshape((2,1,2,3))
-        y = model.constant(np.arange(24)).reshape((2,2,2,3))
-        z = model.constant(np.arange(36)).reshape((2,3,2,3))
-        c = dwave.optimization.symbols.Concatenate((x,y,z), axis=1)
+        x = model.constant(np.arange(12)).reshape((2, 1, 2, 3))
+        y = model.constant(np.arange(24)).reshape((2, 2, 2, 3))
+        z = model.constant(np.arange(36)).reshape((2, 3, 2, 3))
+        c = dwave.optimization.symbols.Concatenate((x, y, z), axis=1)
         model.lock()
         yield c
 
@@ -811,19 +558,19 @@ class TestConcatenate(utils.SymbolTests):
     def test_errors(self):
         model = Model()
         with self.subTest("same number of dimensions"):
-            A = model.constant(np.arange(6)).reshape((1,2,3))
-            B = model.constant(np.arange(24)).reshape((1,2,3,4))
+            A = model.constant(np.arange(6)).reshape((1, 2, 3))
+            B = model.constant(np.arange(24)).reshape((1, 2, 3, 4))
             with self.assertRaisesRegex(
                 ValueError, (r"^all the input arrays must have the same "
                              r"number of dimensions, but the array at index 0 "
                              r"has 3 dimension\(s\) and the array at index 1 "
                              r"has 4 dimension\(s\)")
             ):
-                dwave.optimization.symbols.Concatenate((A,B))
+                dwave.optimization.symbols.Concatenate((A, B))
 
         with self.subTest("array shapes are the same"):
-            A = model.constant(np.arange(6)).reshape((1,2,3))
-            B = model.constant(np.arange(6)).reshape((3,2,1))
+            A = model.constant(np.arange(6)).reshape((1, 2, 3))
+            B = model.constant(np.arange(6)).reshape((3, 2, 1))
             axis = 1
             with self.assertRaisesRegex(
                 ValueError, (r"^all the input array dimensions except for the "
@@ -831,25 +578,25 @@ class TestConcatenate(utils.SymbolTests):
                              r"along dimension 0, the array at index 0 has "
                              r"size 1 and the array at index 1 has size 3")
             ):
-                dwave.optimization.symbols.Concatenate((A,B), axis)
+                dwave.optimization.symbols.Concatenate((A, B), axis)
 
         with self.subTest("axis out of bounds"):
-            A = model.constant(np.arange(6)).reshape((1,2,3))
-            B = model.constant(np.arange(6)).reshape((1,2,3))
+            A = model.constant(np.arange(6)).reshape((1, 2, 3))
+            B = model.constant(np.arange(6)).reshape((1, 2, 3))
             axis = 3
             with self.assertRaisesRegex(
                 ValueError, r"^axis 3 is out of bounds for array of dimension 3"
             ):
-                dwave.optimization.symbols.Concatenate((A,B), axis)
+                dwave.optimization.symbols.Concatenate((A, B), axis)
 
 
-class TestConstant(utils.SymbolTests):
+class TestConstant(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         A = model.constant(np.arange(25, dtype=np.double).reshape((5, 5)))
         B = model.constant([0, 1, 2, 3, 4])
-        C = model.constant(np.arange(25, dtype=np.double))
-        D = model.constant(np.arange(1, 26, dtype=np.double).reshape((5, 5)))
+        model.constant(np.arange(25, dtype=np.double))
+        model.constant(np.arange(1, 26, dtype=np.double).reshape((5, 5)))
         model.lock()
         yield A
         yield B
@@ -959,7 +706,7 @@ class TestConstant(utils.SymbolTests):
             model.constant(np.array([0, 5, np.nan]))
 
 
-class TestCopy(utils.SymbolTests):
+class TestCopy(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         c = model.constant(np.arange(25).reshape(5, 5))
@@ -978,7 +725,7 @@ class TestCopy(utils.SymbolTests):
             np.testing.assert_array_equal(copy.state(), np.arange(25).reshape(5, 5)[::2, 1:4])
 
 
-class TestDisjointBitSetsVariable(utils.SymbolTests):
+class TestDisjointBitSetsVariable(symbol_testing_utils.SymbolTests):
     def test_inequality(self):
         # TODO re-enable this once equality has been fixed
         pass
@@ -1094,7 +841,7 @@ class TestDisjointBitSetsVariable(utils.SymbolTests):
         np.testing.assert_array_equal(ys[2].state(), [0, 0, 0, 0, 1])
 
 
-class TestDisjointListsVariable(utils.SymbolTests):
+class TestDisjointListsVariable(symbol_testing_utils.SymbolTests):
     def test_inequality(self):
         # TODO re-enable this once equality has been fixed
         pass
@@ -1211,7 +958,7 @@ class TestDisjointListsVariable(utils.SymbolTests):
 
 
 # NOTE: Inheriting from BinaryOpTests causes runtime errors so just inheritting from SymbolTests
-class TestDivide(utils.SymbolTests):
+class TestDivide(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         a = model.constant(5)
@@ -1224,8 +971,8 @@ class TestDivide(utils.SymbolTests):
 
     def test_doc_example(self):
         model = Model()
-        i = model.integer(2,lower_bound=1)
-        j = model.integer(2,lower_bound=1)
+        i = model.integer(2, lower_bound=1)
+        j = model.integer(2, lower_bound=1)
         k = i / j
         with model.lock():
             model.states.resize(1)
@@ -1277,7 +1024,7 @@ class TestDivide(utils.SymbolTests):
             a // b
 
 
-class TestIntegerVariable(utils.SymbolTests):
+class TestIntegerVariable(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         x = model.integer(10)
@@ -1294,7 +1041,7 @@ class TestIntegerVariable(utils.SymbolTests):
     def test(self):
         model = Model()
 
-        x = model.integer([10])
+        model.integer([10])
 
     def test_subscript(self):
         model = Model()
@@ -1421,7 +1168,7 @@ class TestIntegerVariable(utils.SymbolTests):
             np.testing.assert_array_equal(x.state(), [0, 0, 0, -1, 0])
 
 
-class TestLessEqual(utils.SymbolTests):
+class TestLessEqual(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         a = model.constant(5)
@@ -1473,7 +1220,7 @@ class TestLessEqual(utils.SymbolTests):
             a <= b
 
 
-class TestListVariable(utils.SymbolTests):
+class TestListVariable(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         x = model.list(10)
@@ -1557,7 +1304,7 @@ class TestListVariable(utils.SymbolTests):
                 x.set_state(0, [0, 1, 2])
 
 
-class TestLogical(utils.UnaryOpTests):
+class TestLogical(symbol_testing_utils.UnaryOpTests):
     def op(self, x):
         return x != 0
 
@@ -1565,7 +1312,7 @@ class TestLogical(utils.UnaryOpTests):
         return logical(x)
 
 
-class TestMax(utils.SymbolTests):
+class TestMax(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         A = model.constant(np.arange(5))
@@ -1595,7 +1342,7 @@ class TestMax(utils.SymbolTests):
         self.assertEqual(b.state(0), 9)
 
 
-class TestMaximum(utils.SymbolTests):
+class TestMaximum(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         A = model.constant(np.arange(5))
@@ -1628,7 +1375,7 @@ class TestMaximum(utils.SymbolTests):
         np.testing.assert_array_equal(m.state(0), np.arange(5, 10))
 
 
-class TestMin(utils.SymbolTests):
+class TestMin(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         A = model.constant(np.arange(5))
@@ -1658,7 +1405,7 @@ class TestMin(utils.SymbolTests):
         self.assertEqual(b.state(0), 5)
 
 
-class TestMinimum(utils.SymbolTests):
+class TestMinimum(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         A = model.constant(np.arange(5))
@@ -1701,7 +1448,7 @@ class TestModule(unittest.TestCase):
             getattr(dwave.optimization.mathematical, name)
 
 
-class TestModulus(utils.BinaryOpTests):
+class TestModulus(symbol_testing_utils.BinaryOpTests):
     def generate_symbols(self):
         model = Model()
         a = model.constant(5)
@@ -1792,7 +1539,7 @@ class TestModulus(utils.BinaryOpTests):
                 np.testing.assert_equal(np_result, result.state(0))
 
 
-class TestMultiply(utils.SymbolTests):
+class TestMultiply(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         a = model.constant(5)
@@ -1829,7 +1576,7 @@ class TestMultiply(utils.SymbolTests):
             a * b
 
 
-class TestNaryAdd(utils.NaryOpTests):
+class TestNaryAdd(symbol_testing_utils.NaryOpTests):
     def op(self, *xs):
         return sum(xs)
 
@@ -1869,7 +1616,7 @@ class TestNaryAdd(utils.NaryOpTests):
             x += b  # after promotion
 
 
-class TestNaryMaximum(utils.NaryOpTests):
+class TestNaryMaximum(symbol_testing_utils.NaryOpTests):
     def op(self, *xs):
         return max(xs)
 
@@ -1880,7 +1627,7 @@ class TestNaryMaximum(utils.NaryOpTests):
         return dwave.optimization.symbols.NaryMaximum
 
 
-class TestNaryMinimum(utils.NaryOpTests):
+class TestNaryMinimum(symbol_testing_utils.NaryOpTests):
     def op(self, *xs):
         return min(xs)
 
@@ -1891,7 +1638,7 @@ class TestNaryMinimum(utils.NaryOpTests):
         return dwave.optimization.symbols.NaryMinimum
 
 
-class TestNaryMultiply(utils.NaryOpTests):
+class TestNaryMultiply(symbol_testing_utils.NaryOpTests):
     def op(self, *xs):
         return math.prod(xs)
 
@@ -1931,12 +1678,12 @@ class TestNaryMultiply(utils.NaryOpTests):
             x *= b  # after promotion
 
 
-class TestNegate(utils.UnaryOpTests):
+class TestNegate(symbol_testing_utils.UnaryOpTests):
     def op(self, x):
         return -x
 
 
-class TestNot(utils.UnaryOpTests):
+class TestNot(symbol_testing_utils.UnaryOpTests):
     def op(self, x):
         return np.logical_not(x)
 
@@ -1944,7 +1691,7 @@ class TestNot(utils.UnaryOpTests):
         return logical_not(x)
 
 
-class TestOr(utils.BinaryOpTests):
+class TestOr(symbol_testing_utils.BinaryOpTests):
     def generate_symbols(self):
         model = Model()
         a = model.constant(1)
@@ -1984,26 +1731,26 @@ class TestOr(utils.BinaryOpTests):
         self.assertEqual(cb.state(0), 1)
 
         with self.assertRaises(TypeError):
-            x = a | b
+            a | b
         with self.assertRaises(AttributeError):
-            x = a.logical_or(b)
+            a.logical_or(b)
 
     def test_array_or(self):
         model = Model()
-        a = model.constant([ 1, 0, 1, 0 ])
-        b = model.constant([ 1, 1, 1, 1 ])
-        c = model.constant([ 0, 0, 0, 0 ])
+        a = model.constant([1, 0, 1, 0])
+        b = model.constant([1, 1, 1, 1])
+        c = model.constant([0, 0, 0, 0])
         ab = logical_or(a, b)
         ac = logical_or(a, c)
 
         model.lock()
         model.states.resize(1)
 
-        np.testing.assert_array_equal(ab.state(0), [ 1, 1, 1, 1 ])
-        np.testing.assert_array_equal(ac.state(0), [ 1, 0, 1, 0 ])
+        np.testing.assert_array_equal(ab.state(0), [1, 1, 1, 1])
+        np.testing.assert_array_equal(ac.state(0), [1, 0, 1, 0])
 
 
-class TestPartialProd(utils.SymbolTests):
+class TestPartialProd(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         A = model.constant(np.arange(8).reshape((2, 2, 2)))
@@ -2034,12 +1781,13 @@ class TestPartialProd(utils.SymbolTests):
         model.lock()
         model.states.resize(1)
         x.set_state(0, [0, 4, 1, 3, 2])
-        np.testing.assert_array_equal(ax.state(0),
-                                      np.prod(np.arange(125).reshape((5, 5, 5))[:, [0, 4, 1, 3, 2], :],
-                                              axis=1))
+        np.testing.assert_array_equal(
+            ax.state(0),
+            np.prod(np.arange(125).reshape((5, 5, 5))[:, [0, 4, 1, 3, 2], :], axis=1)
+        )
 
 
-class TestPartialSum(utils.SymbolTests):
+class TestPartialSum(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         A = model.constant(np.arange(8).reshape((2, 2, 2)))
@@ -2070,12 +1818,13 @@ class TestPartialSum(utils.SymbolTests):
         model.lock()
         model.states.resize(1)
         x.set_state(0, [0, 4, 1, 3, 2])
-        np.testing.assert_array_equal(ax.state(0),
-                                      np.sum(np.arange(125).reshape((5, 5, 5))[:, [0, 4, 1, 3, 2], :],
-                                             axis=1))
+        np.testing.assert_array_equal(
+            ax.state(0),
+            np.sum(np.arange(125).reshape((5, 5, 5))[:, [0, 4, 1, 3, 2], :], axis=1)
+        )
 
 
-class TestPermutation(utils.SymbolTests):
+class TestPermutation(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         A = model.constant(np.arange(25).reshape((5, 5)))
@@ -2101,7 +1850,7 @@ class TestPermutation(utils.SymbolTests):
         self.assertNotIsInstance(b[:, x][x, :], Permutation)
 
 
-class TestProd(utils.SymbolTests):
+class TestProd(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         A = model.constant(np.arange(5))
@@ -2135,7 +1884,7 @@ class TestProd(utils.SymbolTests):
         self.assertEqual(b.state(0), 5*6*7*8*9)
 
 
-class TestPut(utils.SymbolTests):
+class TestPut(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
 
@@ -2190,7 +1939,7 @@ class TestPut(utils.SymbolTests):
             put(array, indices, values)
 
 
-class TestQuadraticModel(utils.SymbolTests):
+class TestQuadraticModel(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         num_variables = 5
         model = Model()
@@ -2279,7 +2028,7 @@ class TestQuadraticModel(utils.SymbolTests):
         self.assertEqual(model.num_nodes(), 2)
 
 
-class TestReshape(utils.SymbolTests):
+class TestReshape(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         A = model.constant(np.arange(12))
@@ -2318,7 +2067,7 @@ class TestReshape(utils.SymbolTests):
             )
 
 
-class TestRint(utils.SymbolTests):
+class TestRint(symbol_testing_utils.SymbolTests):
     rng = np.random.default_rng(1)
 
     def generate_symbols(self):
@@ -2350,7 +2099,7 @@ class TestRint(utils.SymbolTests):
         rint_node_idx6 = rint(model.constant(5.8))
         rint_node_idx8 = rint(model.constant(8))
 
-        arr = model.constant(2 * np.arange(10)) # value at each index is (index * 2)
+        arr = model.constant(2 * np.arange(10))  # value at each index is (index * 2)
         six = arr[rint_node_idx3]
         twelve = arr[rint_node_idx6]
         sixteen = arr[rint_node_idx8]
@@ -2362,12 +2111,12 @@ class TestRint(utils.SymbolTests):
         self.assertEqual(sixteen.state(0), 16)
 
 
-class TestSquare(utils.UnaryOpTests):
+class TestSquare(symbol_testing_utils.UnaryOpTests):
     def op(self, x):
         return x ** 2
 
 
-class TestSquareRoot(utils.SymbolTests):
+class TestSquareRoot(symbol_testing_utils.SymbolTests):
     rng = np.random.default_rng(1)
 
     def generate_symbols(self):
@@ -2400,7 +2149,7 @@ class TestSquareRoot(utils.SymbolTests):
         self.assertRaises(ValueError, sqrt, neg_one)
 
 
-class TestSetVariable(utils.SymbolTests):
+class TestSetVariable(symbol_testing_utils.SymbolTests):
     def generate_symbols(self) -> collections.abc.Iterator[dwave.optimization.symbols.SetVariable]:
         # Typical
         model = Model()
@@ -2497,7 +2246,7 @@ class TestSetVariable(utils.SymbolTests):
         self.assertEqual(model.set(10, max_size=5).state_size(), 5 * 8)
 
 
-class TestSize(utils.SymbolTests):
+class TestSize(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
 
@@ -2534,7 +2283,7 @@ class TestSize(utils.SymbolTests):
             self.assertEqual(length.state(), 1)
 
 
-class TestSubtract(utils.BinaryOpTests):
+class TestSubtract(symbol_testing_utils.BinaryOpTests):
     def generate_symbols(self):
         model = Model()
         a = model.constant(5)
@@ -2574,7 +2323,7 @@ class TestSubtract(utils.BinaryOpTests):
             a - b
 
 
-class TestStack(utils.SymbolTests):
+class TestStack(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         A = model.constant(np.arange(6).reshape((3, 2)))
@@ -2659,7 +2408,7 @@ class TestStack(utils.SymbolTests):
                 stack((A, B), axis=4)
 
 
-class TestSum(utils.SymbolTests):
+class TestSum(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         A = model.constant(np.arange(5))
@@ -2693,7 +2442,7 @@ class TestSum(utils.SymbolTests):
         self.assertEqual(b.state(0), np.arange(5, 10).sum())
 
 
-class TestWhere(utils.SymbolTests):
+class TestWhere(symbol_testing_utils.SymbolTests):
     def generate_symbols(self):
         model = Model()
         condition = model.binary()
@@ -2722,7 +2471,8 @@ class TestWhere(utils.SymbolTests):
             # wrong shape
             dwave.optimization.where(model.binary(3), x, y)
 
-class TestXor(utils.BinaryOpTests):
+
+class TestXor(symbol_testing_utils.BinaryOpTests):
     def generate_symbols(self):
         model = Model()
         a = model.constant(1)
@@ -2762,20 +2512,20 @@ class TestXor(utils.BinaryOpTests):
         self.assertEqual(cb.state(0), 1)
 
         with self.assertRaises(TypeError):
-            x = a ^ b
+            a ^ b
         with self.assertRaises(AttributeError):
-            x = a.logical_xor(b)
+            a.logical_xor(b)
 
     def test_array_xor(self):
         model = Model()
-        a = model.constant([ 1, 0, 1, 0 ])
-        b = model.constant([ 1, 1, 1, 1 ])
-        c = model.constant([ 0, 0, 0, 0 ])
+        a = model.constant([1, 0, 1, 0])
+        b = model.constant([1, 1, 1, 1])
+        c = model.constant([0, 0, 0, 0])
         ab = logical_xor(a, b)
         ac = logical_xor(a, c)
 
         model.lock()
         model.states.resize(1)
 
-        np.testing.assert_array_equal(ab.state(0), [ 0, 1, 0, 1 ])
-        np.testing.assert_array_equal(ac.state(0), [ 1, 0, 1, 0 ])
+        np.testing.assert_array_equal(ab.state(0), [0, 1, 0, 1])
+        np.testing.assert_array_equal(ac.state(0), [1, 0, 1, 0])
